+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.sdp.cryptowalletapp">
 
+    <uses-permission android:name="android.permission.CAMERA" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -10,11 +12,16 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SwiftWallet">
         <activity
-            android:name="com.sdp.cryptowalletapp.GreetingActivity"
-            android:parentActivityName="com.sdp.cryptowalletapp.MainActivity"
-            android:exported="false" />
+            android:name="com.sdp.swiftwallet.QRActivity"
+            android:exported="false"
+            android:parentActivityName="com.sdp.swiftwallet.MainActivity"
+        />
         <activity
-            android:name="com.sdp.cryptowalletapp.MainActivity"
+            android:name="com.sdp.swiftwallet.GreetingActivity"
+            android:exported="false"
+            android:parentActivityName="com.sdp.swiftwallet.MainActivity" />
+        <activity
+            android:name="com.sdp.swiftwallet.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
