--- app/src/main/java/com/sdp/cryptowalletapp/MainActivity.java
@@ -1,27 +1,0 @@
-package com.sdp.cryptowalletapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.EditText;
-
-public class MainActivity extends AppCompatActivity {
-    public static final String EXTRA_MESSAGE="com.sdp.swift-wallet.NAME";
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-    public void sendMessage(View view){
-        //Create a new intent to be able to start a new GreetingActivity
-        Intent intent = new Intent(this, GreetingActivity.class);
-        //Find the EditText object from the main layout by its id;
-        EditText editText = (EditText) findViewById(R.id.mainName);
-        String message = editText.getText().toString();
-        //Put extra parameter to the intent
-        intent.putExtra(EXTRA_MESSAGE, message);
-        startActivity(intent);
-    }
-}
