--- app/src/main/java/com/sdp/cryptowalletapp/common/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.common;
-
-public class Filler {
-}
