--- app/src/main/java/com/sdp/cryptowalletapp/data/remote/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.data.remote;
-
-public class Filler {
-}
