--- app/src/main/java/com/sdp/cryptowalletapp/data/repository/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.data.repository;
-
-public class Filler {
-}
