--- app/src/main/java/com/sdp/cryptowalletapp/domain/model/QRCodeScanner.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.domain.model;
-
-public class QRCodeScanner {
-}
