--- app/src/main/java/com/sdp/cryptowalletapp/domain/repository/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.domain.repository;
-
-public class Filler {
-}
