--- app/src/main/java/com/sdp/cryptowalletapp/domain/use_case/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.domain.use_case;
-
-public class Filler {
-}
