--- app/src/main/java/com/sdp/cryptowalletapp/presentation/login_screen/components/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.presentation.login_screen.components;
-
-public class Filler {
-}
