--- app/src/main/java/com/sdp/cryptowalletapp/presentation/qrcode_converter/components/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.presentation.qrcode_converter.components;
-
-public class Filler {
-}
