--- app/src/main/java/com/sdp/cryptowalletapp/presentation/ui/themes/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.presentation.ui.themes;
-
-public class Filler {
-}
