--- app/src/main/java/com/sdp/cryptowalletapp/presentation/user_profile/components/Filler.java
@@ -1,4 +1,0 @@
-package com.sdp.cryptowalletapp.presentation.user_profile.components;
-
-public class Filler {
-}
