+++ app/src/main/java/com/sdp/swiftwallet/GreetingActivity.java
@@ -1,0 +1,25 @@
+package com.sdp.swiftwallet;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.widget.TextView;
+
+import com.sdp.cryptowalletapp.R;
+
+public class GreetingActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_greeting);
+
+        //Get the intent that started the activity and extract the string  message.
+        Intent intent = getIntent();
+        String message = intent.getStringExtra(MainActivity.EXTRA_MESSAGE);
+
+        TextView textView = findViewById(R.id.mainGreeting);
+        textView.setText(message);
+    }
+}
