+++ app/src/main/java/com/sdp/swiftwallet/common/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.common;
+
+public class Filler {
+}
