+++ app/src/main/java/com/sdp/swiftwallet/data/remote/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.data.remote;
+
+public class Filler {
+}
