+++ app/src/main/java/com/sdp/swiftwallet/data/repository/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.data.repository;
+
+public class Filler {
+}
