+++ app/src/main/java/com/sdp/swiftwallet/domain/model/Login.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.domain.model;
+
+public class Login {
+}
