+++ app/src/main/java/com/sdp/swiftwallet/domain/repository/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.domain.repository;
+
+public class Filler {
+}
