+++ app/src/main/java/com/sdp/swiftwallet/domain/use_case/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.domain.use_case;
+
+public class Filler {
+}
