+++ app/src/main/java/com/sdp/swiftwallet/presentation/login_screen/components/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.presentation.login_screen.components;
+
+public class Filler {
+}
