+++ app/src/main/java/com/sdp/swiftwallet/presentation/qrcode_converter/components/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.presentation.qrcode_converter.components;
+
+public class Filler {
+}
