+++ app/src/main/java/com/sdp/swiftwallet/presentation/ui/themes/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.presentation.ui.themes;
+
+public class Filler {
+}
