+++ app/src/main/java/com/sdp/swiftwallet/presentation/user_profile/components/Filler.java
@@ -1,0 +1,4 @@
+package com.sdp.swiftwallet.presentation.user_profile.components;
+
+public class Filler {
+}
