+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context="com.sdp.swiftwallet.MainActivity">
 
     <EditText
         android:id="@+id/mainName"
@@ -31,6 +31,18 @@
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/mainName"
-        app:layout_constraintVertical_bias="0.214" />
+        app:layout_constraintVertical_bias="0.088" />
+
+    <Button
+        android:id="@+id/qr_goto_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:onClick="startQR"
+        android:text="@string/goto_qr"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/mainButton" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
